+++ NEWS.md
@@ -1,3 +1,8 @@
+# kendallRandomPackage 0.9.0
+
+* Generic function for printing and plotting simulation objects
+* Barrier crossing simulations and visualizations
+
 # kendallRandomPackage 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.